requirements.txt
streamlit
opencv-python
numpy
Pillow
rembg
torch
torchvision
onnxruntime